plan_generator/requirements.txt
aiofiles>=0.8.0
aiohttp>=3.8.0
anthropic
arize-phoenix
asyncio-mqtt
docling
google-genai
mcp-agent
mcp-server-git
nest_asyncio
openai
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
pathlib2
PyPDF2>=2.0.0
reportlab>=3.5.0
streamlit
